split interface and implementation
